Update context folder structure and files
Split the context files into context and provider. Rename variables
and reducer for clarity and consistency.

This fixes the lint issue related to fast refresh, which advised
to move React contexts to separate files.
